requirements.txt
google-play-scraper==1.2.6
pandas==2.2.2
python-dotenv==1.0.0
